fix compiler warnings
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@4565 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/serveractions.h
@@ -80,10 +80,10 @@ struct mapaction : serveraction
             loopv(scl.adminonlymaps)
             {
                 const char *s = scl.adminonlymaps[i], *h = strchr(s, '#'), *m = behindpath(map);
-                int sl = strlen(s);
+                size_t sl = strlen(s);
                 if(h && h != s)
                 {
-                    sl = (int) (h - s);
+                    sl = h - s;
                     if(mode != atoi(h + 1)) continue;
                 }
                 if(sl == strlen(m) && !strncmp(m, scl.adminonlymaps[i], sl)) role = CR_ADMIN;
